Data.Validity

Description

Validity is used to specify additional invariants upon values that are not enforced by the type system.

Let's take an example. Suppose we were to implement a type Prime that represents prime integers.

If you were to completely enforce the invariant that the represented number is a prime, then we could use Natural and only store the index of the given prime in the infinite sequence of prime numbers. This is very safe but also very expensive if we ever want to use the number, because we would have to calculcate all the prime numbers until that index.

Instead we choose to implement Prime by a newtype Prime = Prime Int. Now we have to maintain the invariant that the Int that we use to represent the prime is in fact positive and a prime.

The Validity typeclass allows us to specify this invariant (and enables testing via the genvalidity libraries: https://hackage.haskell.org/package/genvalidity ):

instance Validity Prime where
    validate (Prime n) = check (isPrime n) "The 'Int' is prime."

If certain typeclass invariants exist, you can make these explicit in the validity instance as well. For example, 'Fixed a' is only valid if a has an HasResolution instance, so the correct validity instance is HasResolution a => Validity (Fixed a).

Documentation

class Validity a where

A class of types that have additional invariants defined upon them

Minimal complete definition

Nothing

Methods

validate :: a -> Validation

default validate :: (Generic a, GValidity (Rep a)) => a -> Validation

Helper functions to define validate

trivialValidation :: a -> Validation

Declare any value to be valid in validation

trivialValidation a = seq a mempty

genericValidate :: (Generic a, GValidity (Rep a)) => a -> Validation

check :: Bool -> String -> Validation

Check that a given invariant holds.

The given string should describe the invariant, not the violation.

Example:

check (x < 5) "x is strictly smaller than 5"

instead of

check (x < 5) "x is greater than 5"

declare :: String -> Bool -> Validation

check, but with the arguments flipped

annotate :: Validity a => a -> String -> Validation

Declare a sub-part as a necessary part for validation, and annotate it with a name.

Example:

validate (a, b) =
    mconcat
        [ annotate a "The first element of the tuple"
        , annotate b "The second element of the tuple"
        ]

delve :: Validity a => String -> a -> Validation

annotate, but with the arguments flipped.

decorate :: String -> Validation -> Validation

Decorate a validation with a location

decorateList :: [a] -> (a -> Validation) -> Validation

Decorate a piecewise validation of a list with their location in the list

invalid :: String -> Validation

Construct a trivially invalid Validation

Example:

data Wrong
    = Wrong
    | Fine
    deriving (Show, Eq)

instance Validity Wrong where
    validate w =
        case w of
            Wrong -> invalid "Wrong"
            Fine -> valid

valid :: Validation

Helpers for specific types

Char

validateCharNotUtf16SurrogateCodePoint :: Char -> Validation

isUtf16SurrogateCodePoint :: Char -> Bool

validateCharNotLineSeparator :: Char -> Validation

isLineSeparator :: Char -> Bool

validateStringSingleLine :: String -> Validation

isSingleLine :: String -> Bool

RealFloat (Double)

validateNotNaN :: RealFloat a => a -> Validation

validateNotInfinite :: RealFloat a => a -> Validation

Ratio

validateRatioNotNaN :: Integral a => Ratio a -> Validation

validateRatioNotInfinite :: Integral a => Ratio a -> Validation

validateRatioNormalised :: Integral a => Ratio a -> Validation

Utilities

Utilities for validity checking

isValid :: Validity a => a -> Bool

Check whether a value is valid.

isInvalid :: Validity a => a -> Bool

Check whether a value is not valid.

isInvalid = not . isValid

constructValid :: Validity a => a -> Maybe a

Construct a valid element from an unchecked element

constructValidUnsafe :: (Show a, Validity a) => a -> a

Construct a valid element from an unchecked element, throwing error on invalid elements.

Utilities for validation

newtype Validation

The result of validating a value.

mempty means the value was valid.

This type intentionally doesn't have a Validity instance to make sure you can never accidentally use annotate or delve twice.

Constructors

Validation

Fields

unValidation :: [ValidationChain]

data ValidationChain

Constructors

Violated String

Location String ValidationChain

checkValidity :: Validity a => a -> Either [ValidationChain] a

validate a given value.

This function returns either all the reasons why the given value is invalid, in the form of a list of ValidationChains, or it returns Right with the input value, as evidence that it is valid.

Note: You may want to use prettyValidation instead, if you want to display these ValidationChains to a user.

validationIsValid :: Validation -> Bool

Check if a Validation concerns a valid value.

prettyValidate :: Validity a => a -> Either String a

Validate a given value

This function will return a nice error if the value is invalid. It will return the original value in Right if it was valid, as evidence that it has been validated.

prettyValidation :: Validation -> Maybe String

Render a Validation in a somewhat pretty way.

This function will return Nothing if the Validation concerned a valid value.

Re-exports

class Semigroup a => Monoid a where

The class of monoids (types with an associative binary operation that has an identity). Instances should satisfy the following:

Right identity

x <> mempty = x

Left identity

mempty <> x = x

Associativity

x <> (y <> z) = (x <> y) <> z (Semigroup law)

Concatenation

mconcat = foldr (<>) mempty

The method names refer to the monoid of lists under concatenation, but there are many other instances.

Some types can be viewed as a monoid in more than one way, e.g. both addition and multiplication on numbers. In such cases we often define newtypes and make those instances of Monoid, e.g. Sum and Product.

NOTE: Semigroup is a superclass of Monoid since base-4.11.0.0.

Minimal complete definition

mempty

Methods

mempty :: a

Identity of mappend

>>> "Hello world" <> mempty
"Hello world"

mappend :: a -> a -> a

An associative operation

NOTE: This method is redundant and has the default implementation mappend = (<>) since base-4.11.0.0. Should it be implemented manually, since mappend is a synonym for (<>), it is expected that the two functions are defined the same way. In a future GHC release mappend will be removed from Monoid.

mconcat :: [a] -> a

Fold a list using the monoid.

For most types, the default definition for mconcat will be used, but the function is included in the class definition so that an optimized version can be provided for specific types.

>>> mconcat ["Hello", " ", "Haskell", "!"]
"Hello Haskell!"

class Semigroup a where

The class of semigroups (types with an associative binary operation).

Instances should satisfy the following:

Associativity

x <> (y <> z) = (x <> y) <> z

Since: base-4.9.0.0

Methods

(<>) :: a -> a -> a infixr 6

An associative operation.

>>> [1,2,3] <> [4,5,6]
[1,2,3,4,5,6]